fix: all code reviews notes fixed (except BUG: hot restart returns to the intro screen)
