+++ lib/core/localization/arb/app_en.arb
@@ -1,7 +1,13 @@
 {
     "@@locale": "en",
-    "counterAppBarTitle": "Counter",
-    "@counterAppBarTitle": {
-        "description": "Text shown in the AppBar of the Counter Page"
-    }
+    "getStartedPageMyText": "My",
+    "getStartedPageShopText": "Shop",
+    "getStartedPageLoremIpsumText": "Lorem Ipsum is simply dummy text of the printing and typesetting industry",
+    "getStartedPageGetStartedText": "Get Started",
+    "enterPhonePageWhatIsYourPhoneNumberText": "What Is Your Phone Number?",
+    "enterPhonePageEnterPhoneText": "Please enter your phone number to verify your account",
+    "enterPhonePageUkrainianIdentifierNumsText": "+380 ",
+    "enterPhonePageHintPhoneNumsText": "(99) 999 99 99",
+    "enterPhonePageSendVerificationCodeText": "Send Verification Code",
+    "enterPhonePageSkipText": "Skip"
 }
