+++ lib/core/widgets/send_verification_code_button.dart
@@ -1,14 +1,20 @@
-import 'package:ecommerce_app/core/constants/texts/enter_phone_page_texts.dart';
+import 'package:ecommerce_app/core/localization/localization.dart';
 import 'package:ecommerce_app/core/theme/colors.dart';
+import 'package:ecommerce_app/core/theme/text_styles/enter_phone_page_text_styles.dart';
 import 'package:flutter/material.dart';
 
 class SendVerificationCodeButton extends StatelessWidget {
-  const SendVerificationCodeButton({super.key});
+  const SendVerificationCodeButton({
+    required this.onPressed,
+    super.key,
+  });
+
+  final VoidCallback onPressed;
 
   @override
   Widget build(BuildContext context) {
     return ElevatedButton(
-      onPressed: () {},
+      onPressed: onPressed,
       style: ElevatedButton.styleFrom(
         backgroundColor: AppColors.yellowColor,
         minimumSize: const Size.fromHeight(65),
@@ -16,7 +22,10 @@ class SendVerificationCodeButton extends StatelessWidget {
           borderRadius: BorderRadius.circular(8),
         ),
       ),
-      child: EnterPhonePageTexts.sendVerificationCodeText,
+      child: Text(
+        context.localization.enterPhonePageSendVerificationCodeText,
+        style: EnterPhonePageTextStyles.sendVerificationCodeTextStyle,
+      ),
     );
   }
 }
