+++ lib/src/app/app.dart
@@ -3,12 +3,13 @@ import 'package:ecommerce_app/core/theme/theme.dart';
 import 'package:ecommerce_app/src/app/router/router.dart';
 import 'package:flutter/material.dart';
 
+final AppRouter appRouter = AppRouter();
+
 class App extends StatelessWidget {
   const App({super.key});
 
   @override
   Widget build(BuildContext context) {
-    final appRouter = AppRouter();
     return MaterialApp.router(
       title: 'Ecommerce App',
       theme: AppTheme.themeData,
