+++ lib/src/features/login/presentation/pages/get_started_page/widgets/get_started_button.dart
@@ -1,6 +1,7 @@
 import 'package:auto_route/auto_route.dart';
-import 'package:ecommerce_app/core/constants/texts/get_started_page_texts.dart';
+import 'package:ecommerce_app/core/localization/localization.dart';
 import 'package:ecommerce_app/core/theme/colors.dart';
+import 'package:ecommerce_app/core/theme/text_styles/get_started_page_text_styles.dart';
 import 'package:ecommerce_app/src/app/router/router.dart';
 import 'package:flutter/material.dart';
 
@@ -15,15 +16,18 @@ class GetStartedButton extends StatelessWidget {
       },
       style: ElevatedButton.styleFrom(
         backgroundColor: AppColors.yellowColor,
-        padding: EdgeInsets.symmetric(
+        padding: const EdgeInsets.symmetric(
           vertical: 12,
-          horizontal: MediaQuery.of(context).size.width - 320,
+          horizontal: 75,
         ),
         shape: RoundedRectangleBorder(
           borderRadius: BorderRadius.circular(8),
         ),
       ),
-      child: GetStartedPageTexts.getStartedText,
+      child: Text(
+        context.localization.getStartedPageGetStartedText,
+        style: GetStartedPageTextStyles.getStartedTextStyle,
+      ),
     );
   }
 }
